>fix: Refactor ANCRDT transformation files and move to new location
